Restructure node generation
